+++ src/main/java/com/io/cardtransactions/repository/AccountRepository.java
@@ -4,12 +4,13 @@ import com.io.cardtransactions.domain.Account;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
 import java.util.Optional;
 
 @Repository
-public interface AccountRepository extends JpaRepository<Account, Long> {
+public interface AccountRepository extends JpaRepository<Account, BigInteger> {
 
-    Optional<Account> findById(Long accountId);
+    Optional<Account> findById(BigInteger accountId);
 
     boolean existsByDocumentNumber(String documentNumber);
 }
